feat(cdnjs): Rename manager from `cdn` to `cdnurl`
