+++ lib/manager/cdnurl/extract.ts
