+++ lib/manager/cdnurl/index.ts
