+++ lib/manager/cdnurl/update.ts
