+++ lib/manager/index.ts
@@ -27,7 +27,7 @@ import {
   MANAGER_BUILDKITE,
   MANAGER_BUNDLER,
   MANAGER_CARGO,
-  MANAGER_CDN,
+  MANAGER_CDNURL,
   MANAGER_CIRCLE_CI,
   MANAGER_COMPOSER,
   MANAGER_DEPS_EDN,
@@ -69,7 +69,7 @@ const managerList = [
   MANAGER_BUILDKITE,
   MANAGER_BUNDLER,
   MANAGER_CARGO,
-  MANAGER_CDN,
+  MANAGER_CDNURL,
   MANAGER_CIRCLE_CI,
   MANAGER_COMPOSER,
   MANAGER_DEPS_EDN,
