+++ test/manager/cdnurl/_fixtures/sample.html
